Disable certificate validation
Disable TLS certificate validation when proxying requests to a TLS
enabled backend. All traffic from localshow to the backend servers go
over an encrypted SSH connection, so enabling TLS on the backend will
not add any more security. But if it is enabled, this change should
ensure that things still work, considering there is a slim chance that
the certificate will be valid for the SNI used when proxying requests.

Signed-off-by: Gabriel Adrian Samfira <gsamfira@cloudbasesolutions.com>